Temporary downgrade of wgpu to 0.15.1.
This makes it possible to use `bevy_iced` with the mainline branch of `iced` and still use the most recent official Bevy release, which is still in wgpu 0.15.1.

N.B. Bevy has updated to wgpu 0.16 in mainline, but is likely a while away from an official release. Since iced currently has very few hard dependencies on the new wgpu, it is much simpler to use this temporary downgrade than it is to build against an unstable/mainline Bevy repo.