Add a `Tick` event to the `Event` enum.
Events are currently (and perhaps by design) the only way to trigger a mutation in a widget or canvas program from the outside world. If we want a widget to update every frame, we need an event every frame.

The API is implementation-agnostic as of now; unlike the other event types, it has to be pushed directly onto the `EventQueue` from outside. In the future there might be a self-contained scheduler for this.